cleanup repository, only one is usd
