--- src/Contract/RepositoryInterface.php
@@ -1,22 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Contract;
-
-use JsonSerializable;
-use Symfony\Component\Uid\Uuid;
-
-/**
- * @template TEntity as JsonSerializable
- */
-interface RepositoryInterface
-{
-    public function getTableFile(): string;
-
-    /**
-     * @param array<string, mixed> $row
-     * @return TEntity
-     */
-    public function createEntity(Uuid $uuid, array $row): JsonSerializable;
-}
